fix warning

+++ src/vm/wren_utils.c
@@ -201,7 +201,7 @@ uint32_t wrenValidateIndex(uint32_t count, int64_t value)
   if (value < 0) value = count + value;
 
   // Check bounds.
-  if (value >= 0 && value < count) return value;
+  if (value >= 0 && value < count) return (uint32_t)value;
 
   return UINT32_MAX;
 }
